SOme small refactoring, creating the hire mechanism. units common value for random adding.
Added "Common" value to define the chanse value to be randomly added. a common of 10 means that a roll must be equal or minor to 10 to be added.
now units get common value to define how common are when are added randomly.
Buildings also have a common value.
Start the creation of hire for random units.
some small refactoring.